Classes in separate files
- Also fixed references in all other files
