+++ examples/dithering3_2/index.js
@@ -1,7 +1,8 @@
 import vert from './vert.js';
 import compute from './compute.js';
 import frag from './frag.js';
-import { RenderPass, ShaderType } from '../../src/absulit.points.module.js';
+import RenderPass from './../../src/RenderPass.js';
+import ShaderType from './../../src/ShaderType.js';
 const dithering3 = {
     renderPasses: [
         new RenderPass(vert, frag, compute, 800, 800)
