+++ examples/renderpasses2/index.js
@@ -1,8 +1,8 @@
 import vert1 from './renderpass1/vert.js';
 import frag1 from './renderpass1/frag.js';
 
-import { RenderPass } from '../../src/absulit.points.module.js';
-import { RenderPasses } from "../../src/RenderPasses.js";
+import RenderPass from '../../src/RenderPass.js';
+import RenderPasses from './../../src/RenderPasses.js';
 
 const renderpasses1 = {
     /**
