+++ src/ShaderType.js
@@ -1,0 +1,7 @@
+'use strict';
+
+export default class ShaderType {
+    static VERTEX = 1;
+    static COMPUTE = 2;
+    static FRAGMENT = 3;
+}
