+++ src/UniformKeys.js
@@ -1,0 +1,15 @@
+'use strict';
+
+export default class UniformKeys {
+    static TIME = 'time';
+    static EPOCH = 'epoch';
+    static SCREEN_WIDTH = 'screenWidth';
+    static SCREEN_HEIGHT = 'screenHeight';
+    static MOUSE_X = 'mouseX';
+    static MOUSE_Y = 'mouseY';
+    static MOUSE_CLICK = 'mouseClick';
+    static MOUSE_DOWN = 'mouseDown';
+    static MOUSE_WHEEL = 'mouseWheel';
+    static MOUSE_DELTA_X = 'mouseDeltaX';
+    static MOUSE_DELTA_Y = 'mouseDeltaY';
+}
