+++ src/VertexBufferInfo.js
@@ -1,0 +1,39 @@
+'use strict';
+
+export default class VertexBufferInfo {
+    /**
+     * Along with the vertexArray it calculates some info like offsets required for the pipeline.
+     * @param {Float32Array} vertexArray array with vertex, color and uv data
+     * @param {Number} triangleDataLength how many items does a triangle row has in vertexArray
+     * @param {Number} vertexOffset index where the vertex data starts in a row of `triangleDataLength` items
+     * @param {Number} colorOffset index where the color data starts in a row of `triangleDataLength` items
+     * @param {Number} uvOffset index where the uv data starts in a row of `triangleDataLength` items
+     */
+    constructor(vertexArray, triangleDataLength = 10, vertexOffset = 0, colorOffset = 4, uvOffset = 8) {
+        this._vertexSize = vertexArray.BYTES_PER_ELEMENT * triangleDataLength; // Byte size of ONE triangle data (vertex, color, uv). (one row)
+        this._vertexOffset = vertexArray.BYTES_PER_ELEMENT * vertexOffset;
+        this._colorOffset = vertexArray.BYTES_PER_ELEMENT * colorOffset; // Byte offset of triangle vertex color attribute.
+        this._uvOffset = vertexArray.BYTES_PER_ELEMENT * uvOffset;
+        this._vertexCount = vertexArray.byteLength / this._vertexSize;
+    }
+
+    get vertexSize() {
+        return this._vertexSize;
+    }
+
+    get vertexOffset() {
+        return this._vertexOffset;
+    }
+
+    get colorOffset() {
+        return this._colorOffset;
+    }
+
+    get uvOffset() {
+        return this._uvOffset;
+    }
+
+    get vertexCount() {
+        return this._vertexCount;
+    }
+}
