+++ src/absulit.points.module.js
@@ -1,189 +1,11 @@
 'use strict';
+import UniformKeys from './UniformKeys.js';
+import VertexBufferInfo from './VertexBufferInfo.js';
+import ShaderType from './ShaderType.js';
 import Coordinate from './coordinate.js';
 import RGBAColor from './color.js';
 import defaultStructs from './core/defaultStructs.js';
 import { defaultVertexBody } from './core/defaultFunctions.js';
-
-export class ShaderType {
-    static VERTEX = 1;
-    static COMPUTE = 2;
-    static FRAGMENT = 3;
-}
-
-class UniformKeys {
-    static TIME = 'time';
-    static EPOCH = 'epoch';
-    static SCREEN_WIDTH = 'screenWidth';
-    static SCREEN_HEIGHT = 'screenHeight';
-    static MOUSE_X = 'mouseX';
-    static MOUSE_Y = 'mouseY';
-    static MOUSE_CLICK = 'mouseClick';
-    static MOUSE_DOWN = 'mouseDown';
-    static MOUSE_WHEEL = 'mouseWheel';
-    static MOUSE_DELTA_X = 'mouseDeltaX';
-    static MOUSE_DELTA_Y = 'mouseDeltaY';
-}
-
-export class RenderPass {
-    /**
-     * A collection of Vertex, Compute and Fragment shaders that represent a RenderPass.
-     * This is useful for PostProcessing.
-     * @param {String} vertexShader  WGSL Vertex Shader in a String.
-     * @param {String} fragmentShader  WGSL Fragment Shader in a String.
-     * @param {String} computeShader  WGSL Compute Shader in a String.
-     */
-    constructor(vertexShader, fragmentShader, computeShader, workgroupCountX, workgroupCountY, workgroupCountZ) {
-        this._vertexShader = vertexShader;
-        this._computeShader = computeShader;
-        this._fragmentShader = fragmentShader;
-
-        this._computePipeline = null;
-        this._renderPipeline = null;
-
-        this._computeBindGroup = null;
-        this._uniformBindGroup = null;
-
-        this._internal = false;
-
-        this._compiledShaders = {
-            vertex: '',
-            compute: '',
-            fragment: '',
-        }
-
-        this._hasComputeShader = !!this._computeShader;
-        this._hasVertexShader = !!this._vertexShader;
-        this._hasFragmentShader = !!this._fragmentShader;
-
-        this._hasVertexAndFragmentShader = this._hasVertexShader && this._hasFragmentShader;
-
-        this._workgroupCountX = workgroupCountX || 8;
-        this._workgroupCountY = workgroupCountY || 8;
-        this._workgroupCountZ = workgroupCountZ || 1;
-    }
-
-    get internal() {
-        return this._internal;
-    }
-
-    set internal(value) {
-        this._internal = value;
-    }
-
-    get vertexShader() {
-        return this._vertexShader;
-    }
-
-    get computeShader() {
-        return this._computeShader;
-    }
-
-    get fragmentShader() {
-        return this._fragmentShader;
-    }
-
-    set computePipeline(value) {
-        this._computePipeline = value;
-    }
-
-    get computePipeline() {
-        return this._computePipeline;
-    }
-
-    set renderPipeline(value) {
-        this._renderPipeline = value;
-    }
-
-    get renderPipeline() {
-        return this._renderPipeline;
-    }
-
-    set computeBindGroup(value) {
-        this._computeBindGroup = value;
-    }
-
-    get computeBindGroup() {
-        return this._computeBindGroup;
-    }
-
-    set uniformBindGroup(value) {
-        this._uniformBindGroup = value;
-    }
-
-    get uniformBindGroup() {
-        return this._uniformBindGroup;
-    }
-
-    get compiledShaders() {
-        return this._compiledShaders;
-    }
-
-    get hasComputeShader() {
-        return this._hasComputeShader;
-    }
-
-    get hasVertexShader() {
-        return this._hasVertexShader;
-    }
-
-    get hasFragmentShader() {
-        return this._hasFragmentShader;
-    }
-
-    get hasVertexAndFragmentShader() {
-        return this._hasVertexAndFragmentShader;
-    }
-
-    get workgroupCountX() {
-        return this._workgroupCountX;
-    }
-
-    get workgroupCountY() {
-        return this._workgroupCountY;
-    }
-
-    get workgroupCountZ() {
-        return this._workgroupCountZ;
-    }
-}
-
-export class VertexBufferInfo {
-    /**
-     * Along with the vertexArray it calculates some info like offsets required for the pipeline.
-     * @param {Float32Array} vertexArray array with vertex, color and uv data
-     * @param {Number} triangleDataLength how many items does a triangle row has in vertexArray
-     * @param {Number} vertexOffset index where the vertex data starts in a row of `triangleDataLength` items
-     * @param {Number} colorOffset index where the color data starts in a row of `triangleDataLength` items
-     * @param {Number} uvOffset index where the uv data starts in a row of `triangleDataLength` items
-     */
-    constructor(vertexArray, triangleDataLength = 10, vertexOffset = 0, colorOffset = 4, uvOffset = 8) {
-        this._vertexSize = vertexArray.BYTES_PER_ELEMENT * triangleDataLength; // Byte size of ONE triangle data (vertex, color, uv). (one row)
-        this._vertexOffset = vertexArray.BYTES_PER_ELEMENT * vertexOffset;
-        this._colorOffset = vertexArray.BYTES_PER_ELEMENT * colorOffset; // Byte offset of triangle vertex color attribute.
-        this._uvOffset = vertexArray.BYTES_PER_ELEMENT * uvOffset;
-        this._vertexCount = vertexArray.byteLength / this._vertexSize;
-    }
-
-    get vertexSize() {
-        return this._vertexSize
-    }
-
-    get vertexOffset() {
-        return this._vertexOffset;
-    }
-
-    get colorOffset() {
-        return this._colorOffset;
-    }
-
-    get uvOffset() {
-        return this._uvOffset;
-    }
-
-    get vertexCount() {
-        return this._vertexCount;
-    }
-}
 
 export default class Points {
     constructor(canvasId) {
