+++ src/core/RenderPasses/filmgrain/index.js
@@ -1,6 +1,6 @@
 import vertexShader from './vert.js';
 import fragmentShader from './frag.js';
-import { ShaderType } from '../../../absulit.points.module.js';
+import ShaderType from '../../../absulit.points.module.js';
 
 const filmgrain = {
     vertexShader,
